Set eps type to krylov-schur in setupEPS
When eps.setKrylovSchurSubintervals() is called for the
first time I get the following error:
[2] Cannot locate function EPSKrylovSchurGetPartitions_C in object
Following Jose's suggestion fixes the problem.

Here is Jose's explanation:
In SLEPc and PETSc, methods that are specific of a solver type, such as EPSKrylovSchurSet/GetPartitions work as follows:
1) If the object type has been set (to KrylovSchur in this case) then both Set and Get work.
2) If the object type has not been set, then the Set operation does not fail (it has no effect) and the Get operation yields an error because the object does not have the correct type.

Your error should go away if you add this:
    eps.setType(SLEPc.EPS.Type.KRYLOVSCHUR)
before any KrylovSchur-specific call.

Now the question is why is it complaining if all calls are Set and not Get? The reason is that within slepc4py one of the Set operations (setKrylovSchurSubintervals) is calling a Get operation (EPSKrylovSchurGetPartitions).