Fix the bug for usesips=False
Jose reported that PSCf was not working without sips.
It was broken while adding support for local update of eps operators.
Now PSCF works with or without sips using euither local or glabal update
of eps operators.
However, when sips is used only local update of eps is possible, and
it is the best option for performance.